Gosh, I have no idea if this one is even better than the last one.
Added the scrollbar stuff into MapView instead of it being part of the demo. Consequently split MapView into a parent class and subclass. Talk about going in circles.

Probably a bunch of other nonsense I forgot about.

In any case, it's OK, I might (re)add scrolling this week but I'm happy with it as is.