Added my Spell Index method and associated files
The files in this directory are associated with the spell index draft.
It is implemented here as a method of a SpellBook class.
When passed a character, it returns an array of spells that character
may attempt, or a list containing only the empty spell if the character
is not a wizard.

See src\spellindex\readme.txt for more details.