Removed "Movement Calculator" Run Environment
Since the movement calculations are now done in a method of of
src/model/sshexmap/MapHex.java, this configuration no longer pointed to
an actual class, so I removed it. There is a Main.java in the ssterrain
package (although it's not in src/model/ssterrain) that demonstrates it
in action.