MapDemoScenario reads in Scenario 0 to map
The unit populator in the scenario config reader can now take simple
setups, such as player-owned units it knows the name ofwhen the nation
names match the name of a Race it knows about. Neutrals are not yet
handled and these units do not go on a specific map location yet. We
will slowly build up Scenario 0.