Changed the way Game loads JavaFX files.
Game now uses an FXMLoader instance to load the HUD. By using
an instance rather than the static FXMLoader methods, we can
get a reference to the actual hudController instance used
to update the hud.

We're storing this reference so that the network Conductor object
can initiate changes to the GUI when certain network events occur.