Got Netbeans to test finally Chat is almost working, just need to rewrite one method and several calls to said method. After I do that, I can implement client to client chat as well.
The issue lies with how the NetworkServer is passing along the broadcast. It recieves its arguments as var-args, which then gets passed into another varargs. var-args is just "syntatic sugar" for an array, so basically the first element is getting an array of one or more items.

If enforce prepacking of Lists for everything but the most external method, then things should be fine and dandy!